+++ account/src/main/java/com/ray/service/account/model/Account.java
@@ -13,7 +13,7 @@ public class Account implements Serializable {
     public static final int STATUS_EXPIRED = 1;
     public static final int STATUS_LOCKED = 2;
     public static final int STATUS_DISABLED = 3;
-    private int uid = 0;
+    private long uid = 0;
     private String username;
     private String password;
     private int status = STATUS_NORMAL;
@@ -22,11 +22,11 @@ public class Account implements Serializable {
     private long lastlogin;
     private List<Authority> roles;
 
-    public int getUid() {
+    public long getUid() {
         return uid;
     }
 
-    public void setUid(int uid) {
+    public void setUid(long uid) {
         this.uid = uid;
     }
